# [Enhancement] Move cte inline to memo phase 1