--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/Optimizer.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/Optimizer.java
@@ -135,6 +135,7 @@ public class Optimizer {
         if (!sessionVariable.isDisableJoinReorder()
                 && Utils.countInnerJoinNodeSize(tree) < sessionVariable.getCboMaxReorderNode()) {
             if (Utils.countInnerJoinNodeSize(tree) > sessionVariable.getCboMaxReorderNodeUseExhaustive()) {
+                CTEUtils.collectCteStatistics(memo, context);
                 new ReorderJoinRule().transform(tree, context);
                 context.getRuleSet().addJoinCommutativityWithOutInnerRule();
             } else {
@@ -192,7 +193,7 @@ public class Optimizer {
         CTEUtils.collectCteOperatorsWithoutCosts(memo, context);
         // inline CTE if consume use once
         while (cteContext.hasInlineCTE()) {
-            ruleRewriteOnlyOnce(memo, rootTaskContext, RuleSetType.INLINE_ONE_CTE);
+            ruleRewriteOnlyOnce(memo, rootTaskContext, RuleSetType.INLINE_CTE);
             CTEUtils.collectCteOperatorsWithoutCosts(memo, context);
         }
         cleanUpMemoGroup(memo);
@@ -279,13 +280,13 @@ public class Optimizer {
 
         // compute CTE inline by costs
         if (cteContext.needOptimizeCTE()) {
-            CTEUtils.collectCteOperators(memo, context);
+            CTEUtils.collectCteOperatorsWithoutCosts(memo, context);
         }
 
-        // compute CTE inline by costs
-        while (cteContext.needOptimizeCTE() && cteContext.hasInlineCTE()) {
+        // inline CTE if consume use once
+        while (cteContext.hasInlineCTE()) {
             ruleRewriteOnlyOnce(memo, rootTaskContext, RuleSetType.INLINE_CTE);
-            CTEUtils.collectCteOperators(memo, context);
+            CTEUtils.collectCteOperatorsWithoutCosts(memo, context);
         }
 
         ruleRewriteIterative(memo, rootTaskContext, new MergeTwoProjectRule());