--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/PropertyDeriverBase.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/PropertyDeriverBase.java
@@ -133,7 +133,4 @@ public abstract class PropertyDeriverBase<R, C> extends OperatorVisitor<R, C> {
         return new PhysicalPropertySet(distributionProperty);
     }
 
-    protected PhysicalPropertySet createBroadcastPropertySet() {
-        return new PhysicalPropertySet(new DistributionProperty(DistributionSpec.createReplicatedDistributionSpec()));
-    }
 }