--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/base/CTEProperty.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/base/CTEProperty.java
@@ -1,4 +1,4 @@
-// This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Limited.
+// This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
 
 package com.starrocks.sql.optimizer.base;