--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/RuleSetType.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/RuleSetType.java
@@ -17,5 +17,4 @@ public enum RuleSetType {
     PRUNE_PROJECT,
     COLLECT_CTE,
     INLINE_CTE,
-    INLINE_ONE_CTE,
 }