… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/base/PhysicalPropertySet.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/cost/CostModel.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/logical/LogicalCTEConsumeOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/physical/PhysicalCTEAnchorOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/physical/PhysicalNoCTEOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/RuleSet.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/RuleType.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/implementation/CTEAnchorImplementationRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/implementation/CTEAnchorToNoCTEImplementationRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/implementation/CTEConsumeInlineImplementationRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/implementation/CTEConsumerReuseImplementationRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/InlineOneCTEConsumeRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/statistics/StatisticsCalculator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/task/EnforceAndCostTask.java
… fe/fe-core/src/test/java/com/starrocks/sql/optimizer/OperatorStrings.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/CTEPlanTest.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/ReplayFromDumpTest.java