## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] refactor
- [ ] others

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->

Move cte inline rule to memo phase, we will enforce all inline or all reuse cte

1. add CTENoOP implements
2. add CTEProperty in Enforce phase
3. add CTE costs model


Opimizer Performance:

|  Query  | Before | After | 
|  ----  | ----  | ---- |
|  multi-nest view query with cte(8+) #9347  | 300ms | 300ms |
| 20 iterate cte + 60 refs  | 100ms | - |
| 12 iterate cte + 36 refs  | - | 2224ms  |
| 11 iterate cte + 33 refs  | - | 700ms  |
| 10 iterate cte + 30 refs  | - | 280ms  |
| 8 nested cte + 3 * 2^8 refs  | 4081ms | 7522ms  |
| 7 nested cte + 3 * 2^7 refs  | 1481ms | 2250ms  |
| 6 nested cte + 3 * 2^6 refs  | 451ms | 600ms  |

Of course, the cost of these test is in RuleWrite phase, not in memo phase.It's will take more time if we test so many joins or complex SQL

BTW, we need optimize the cost of memo merge 


SQLs
```
# iterater cte
with x1 as (select * from t0),
     x2 as (select * from t0),
     x3 as (select * from t0),
     x4 as (select * from t0),
     x5 as (select * from t0),
     x6 as (select * from t0),                    
     x7 as (select * from t0),
     x8 as (select * from t0),          
     x9 as (select * from t0),          
     x10 as (select * from t0),
     x11 as (select * from t0)     
select * from x1 union all
select * from x1 union all
select * from x1 union all
select * from x2 union all
select * from x2 union all
select * from x2 union all
select * from x3 union all
select * from x3 union all
select * from x3 union all
select * from x4 union all
select * from x4 union all
select * from x4 union all
select * from x5 union all
select * from x5 union all
select * from x5 union all
select * from x6 union all
select * from x6 union all
select * from x6 union all
select * from x7 union all
select * from x7 union all
select * from x7 union all
select * from x8 union all
select * from x8 union all
select * from x8 union all
select * from x9 union all
select * from x9 union all
select * from x9 union all
select * from x10 union all
select * from x10 union all
select * from x10 union all
select * from x11 union all
select * from x11 union all
select * from x11;

# nest cte
explain 
with x1 as (select * from t0),
     x2 as (select * from x1 union all 
            select * from x1),
     x3 as (select * from x2 union all 
            select * from x2),
     x4 as (select * from x3 union all 
            select * from x3),
     x5 as (select * from x4 union all 
            select * from x4),
     x6 as (select * from x5 union all 
            select * from x5),
     x7 as (select * from x6 union all 
            select * from x6)
select * from x1 union all
select * from x1 union all
select * from x1 union all
select * from x2 union all
select * from x2 union all
select * from x2 union all
select * from x3 union all
select * from x3 union all
select * from x3 union all
select * from x4 union all
select * from x4 union all
select * from x4 union all
select * from x5 union all
select * from x5 union all
select * from x5 union all
select * from x6 union all
select * from x6 union all
select * from x6 union all
select * from x7 union all
select * from x7 union all
select * from x7 ;


```