@wanpengfei-git 2022-08-08
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 173 / 188 (92.02%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/logical/LogicalCTEConsumeOperator.java|0|1|00.00%|[84]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/InlineCTEConsumeRule.java|0|4|00.00%|[14, 17, 21, 22]|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/physical/PhysicalCTEAnchorOperator.java|1|2|50.00%|[29]|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/physical/PhysicalNoCTEOperator.java|4|5|80.00%|[43]|
|:large_blue_circle:|com/starrocks/sql/optimizer/base/CTEProperty.java|24|29|82.76%|[53, 58, 59, 68, 81]|
|:large_blue_circle:|com/starrocks/sql/optimizer/statistics/StatisticsCalculator.java|12|13|92.31%|[1364]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/EnforceAndCostTask.java|14|15|93.33%|[215]|
|:large_blue_circle:|com/starrocks/sql/optimizer/base/PhysicalPropertySet.java|14|15|93.33%|[97]|
|:large_blue_circle:|com/starrocks/sql/optimizer/RequiredPropertyDeriver.java|26|26|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/Optimizer.java|5|5|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/implementation/CTEConsumerReuseImplementationRule.java|1|1|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/implementation/CTEAnchorToNoCTEImplementationRule.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/implementation/CTEConsumeInlineImplementationRule.java|8|8|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/implementation/CTEAnchorImplementationRule.java|3|3|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/ChildOutputPropertyGuarantor.java|11|11|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/RuleType.java|2|2|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/OutputPropertyDeriver.java|24|24|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/InlineOneCTEConsumeRule.java|6|6|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/CTEContext.java|8|8|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/CTEUtils.java|4|4|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/cost/CostModel.java|4|4|100.00%|[]|