@kangkaisen 2022-08-08
Need the optimizer itself performance test, we need to know the optimizer will slow how much time.