@Seaven 2022-08-15
run starrocks_fe_unittest